Shared/requirements.txt
grpcio==1.74.0
loguru==0.7.3
protobuf==6.31.1
